1,修改了orchestration-service.cds，修改了BusinessScenarios的exclude字段，添加embeddings_ai。 2.修改了Bot，添加了getLocale方法定义 3.修改了ChatBot,CodingBot,FunctionCallingBot，添加了Locale属性，调用promptService.getPrompt方法修改了传参改成this 4.修改了ChatBot，修改了chat方法，添加了获取RAG内容的逻辑。将用户消息存表逻辑从BotServiceImpl移动到ChatBot.chat方法。 5.修改了PromptService，修改了getPrompt方法定义传参，改成Bot对象，添加了getRagAsPrompts的方法 6.修改了PromptServiceImpl，添加了RAG的逻辑。 7.修改了RAGExtrator，修改了threshold参数的类型改成double 8.修改了GenericCqnService，添加了createAndInsertBotMessage方法，增加传参ragContent 9.修改了RAGCDSViewExtratorImpl,RAGJoinConditionExtratorImpl,RAGViewFieldExtratorImpl，修改了threshold的传参类型

+++ srv/src/main/java/customer/ai2code/service/execution/RAGExtractor.java
@@ -5,5 +5,5 @@ import java.util.Locale;
 import cds.gen.configservice.PromptTexts;
 
 public interface RAGExtractor {
-    public String extract(String ragSource, int ragTopK, String query, Locale language, int threshold);
+    public String extract(String ragSource, int ragTopK, String query, Locale language, double threshold);
 }
